Move AccuWeatherConfigEntry/Data to coordinator module
