+++ homeassistant/components/accuweather/__init__.py
@@ -2,21 +2,21 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 import logging
 
 from accuweather import AccuWeather
 
 from homeassistant.components.sensor import DOMAIN as SENSOR_PLATFORM
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_API_KEY, CONF_NAME, Platform
+from homeassistant.const import CONF_API_KEY, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
-from .const import DOMAIN, UPDATE_INTERVAL_DAILY_FORECAST, UPDATE_INTERVAL_OBSERVATION
+from .const import DOMAIN
 from .coordinator import (
+    AccuWeatherConfigEntry,
     AccuWeatherDailyForecastDataUpdateCoordinator,
+    AccuWeatherData,
     AccuWeatherObservationDataUpdateCoordinator,
 )
 
@@ -25,21 +25,9 @@ _LOGGER = logging.getLogger(__name__)
 PLATFORMS = [Platform.SENSOR, Platform.WEATHER]
 
 
-@dataclass
-class AccuWeatherData:
-    """Data for AccuWeather integration."""
-
-    coordinator_observation: AccuWeatherObservationDataUpdateCoordinator
-    coordinator_daily_forecast: AccuWeatherDailyForecastDataUpdateCoordinator
-
-
-type AccuWeatherConfigEntry = ConfigEntry[AccuWeatherData]
-
-
 async def async_setup_entry(hass: HomeAssistant, entry: AccuWeatherConfigEntry) -> bool:
     """Set up AccuWeather as config entry."""
     api_key: str = entry.data[CONF_API_KEY]
-    name: str = entry.data[CONF_NAME]
 
     location_key = entry.unique_id
 
@@ -49,19 +37,11 @@ async def async_setup_entry(hass: HomeAssistant, entry: AccuWeatherConfigEntry) 
     accuweather = AccuWeather(api_key, websession, location_key=location_key)
 
     coordinator_observation = AccuWeatherObservationDataUpdateCoordinator(
-        hass,
-        accuweather,
-        name,
-        "observation",
-        UPDATE_INTERVAL_OBSERVATION,
+        hass, entry, accuweather
     )
 
     coordinator_daily_forecast = AccuWeatherDailyForecastDataUpdateCoordinator(
-        hass,
-        accuweather,
-        name,
-        "daily forecast",
-        UPDATE_INTERVAL_DAILY_FORECAST,
+        hass, entry, accuweather
     )
 
     await coordinator_observation.async_config_entry_first_refresh()
