+++ homeassistant/components/accuweather/coordinator.py
@@ -1,13 +1,17 @@
 """The AccuWeather coordinator."""
 
+from __future__ import annotations
+
 from asyncio import timeout
-from datetime import timedelta
+from dataclasses import dataclass
 import logging
 from typing import TYPE_CHECKING, Any
 
 from accuweather import AccuWeather, ApiError, InvalidApiKeyError, RequestsExceededError
 from aiohttp.client_exceptions import ClientConnectorError
 
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import CONF_NAME
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.update_coordinator import (
@@ -16,11 +20,27 @@ from homeassistant.helpers.update_coordinator import (
     UpdateFailed,
 )
 
-from .const import DOMAIN, MANUFACTURER
+from .const import (
+    DOMAIN,
+    MANUFACTURER,
+    UPDATE_INTERVAL_DAILY_FORECAST,
+    UPDATE_INTERVAL_OBSERVATION,
+)
 
 EXCEPTIONS = (ApiError, ClientConnectorError, InvalidApiKeyError, RequestsExceededError)
 
 _LOGGER = logging.getLogger(__name__)
+
+
+@dataclass
+class AccuWeatherData:
+    """Data for AccuWeather integration."""
+
+    coordinator_observation: AccuWeatherObservationDataUpdateCoordinator
+    coordinator_daily_forecast: AccuWeatherDailyForecastDataUpdateCoordinator
+
+
+type AccuWeatherConfigEntry = ConfigEntry[AccuWeatherData]
 
 
 class AccuWeatherObservationDataUpdateCoordinator(
@@ -31,10 +51,8 @@ class AccuWeatherObservationDataUpdateCoordinator(
     def __init__(
         self,
         hass: HomeAssistant,
+        config_entry: AccuWeatherConfigEntry,
         accuweather: AccuWeather,
-        name: str,
-        coordinator_type: str,
-        update_interval: timedelta,
     ) -> None:
         """Initialize."""
         self.accuweather = accuweather
@@ -42,14 +60,14 @@ class AccuWeatherObservationDataUpdateCoordinator(
 
         if TYPE_CHECKING:
             assert self.location_key is not None
-
+        name: str = config_entry.data[CONF_NAME]
         self.device_info = _get_device_info(self.location_key, name)
 
         super().__init__(
             hass,
             _LOGGER,
-            name=f"{name} ({coordinator_type})",
-            update_interval=update_interval,
+            name=f"{name} (observation)",
+            update_interval=UPDATE_INTERVAL_OBSERVATION,
         )
 
     async def _async_update_data(self) -> dict[str, Any]:
@@ -73,10 +91,8 @@ class AccuWeatherDailyForecastDataUpdateCoordinator(
     def __init__(
         self,
         hass: HomeAssistant,
+        config_entry: AccuWeatherConfigEntry,
         accuweather: AccuWeather,
-        name: str,
-        coordinator_type: str,
-        update_interval: timedelta,
     ) -> None:
         """Initialize."""
         self.accuweather = accuweather
@@ -85,13 +101,14 @@ class AccuWeatherDailyForecastDataUpdateCoordinator(
         if TYPE_CHECKING:
             assert self.location_key is not None
 
+        name: str = config_entry.data[CONF_NAME]
         self.device_info = _get_device_info(self.location_key, name)
 
         super().__init__(
             hass,
             _LOGGER,
-            name=f"{name} ({coordinator_type})",
-            update_interval=update_interval,
+            name=f"{name} (daily forecast)",
+            update_interval=UPDATE_INTERVAL_DAILY_FORECAST,
         )
 
     async def _async_update_data(self) -> list[dict[str, Any]]:
